moving migrations to module
