+++ lib/migrate/index.js
@@ -1,0 +1,4 @@
+const { runCLIAction } = require('./migrate');
+module.exports = {
+  runCLIAction,
+};
