+++ lib/migrate/loadEnvs.js
@@ -7,7 +7,7 @@ const path = require('path');
 
 function loadEnvSettings() {
   // FIXME: hardcoded filenames
-  const envFilePaths = ['.env.local', '.env.development.local', '.env.development'];
+  const envFilePaths = ['.env.development.local', '.env.development', '.env.local'];
 
   // Hack: find the first one that exist
   const curDir = process.cwd();
@@ -18,8 +18,8 @@ function loadEnvSettings() {
   }
 
   try {
+    console.log(`Loading environment settings from ${fp}`);
     dotenv.config({ path: fp });
-    console.log(`loaded environment file: ${fp}`);
   } catch (err) {
     throw new Error(`error: failed to load environment file '${fp}': ${err}`);
   }
