+++ lib/migrate/migrate.js
@@ -21,6 +21,23 @@ const actions = {
   status: migrationsStatus,
 };
 
+async function runCLIAction() {
+  const [, , direction] = process.argv;
+  const action = actions[direction];
+
+  if (!action) {
+    console.error(`Expected ${Object.keys(actions).join(', ')} but got ${direction}`);
+    process.exit(1);
+  }
+
+  const rows = await ensureMigrationsTable();
+  const files = fs.readdirSync(migrationsDir).sort((a, b) => (a > b ? 1 : -1));
+  await action({ rows, files });
+
+  console.log(`migration complete`);
+  process.exit(0);
+}
+
 function getMigrationsDir() {
   let dir = process.env.MIGRATIONS_DIR || './migrations/';
   if (!dir.endsWith('/')) {
@@ -138,21 +155,6 @@ async function ensureMigrationsTable() {
   return await getMigrationRows();
 }
 
-async function runCLIAction() {
-  const [, , direction] = process.argv;
-  const action = actions[direction];
-
-  if (!action) {
-    console.error(`Expected ${Object.keys(actions).join(', ')} but got ${direction}`);
-    process.exit(1);
-  }
-
-  const rows = await ensureMigrationsTable();
-  const files = fs.readdirSync(migrationsDir).sort((a, b) => (a > b ? 1 : -1));
-  await action({ rows, files });
-
-  console.log(`migration complete`);
-  process.exit(0);
-}
-
-runCLIAction();
+module.exports = {
+  runCLIAction,
+};
