+++ lib/migrate/migration-template.js
