+++ script/migrate.js
@@ -1,0 +1,4 @@
+// Run migrations from the CLI
+const { runCLIAction } = require('../lib/migrate');
+
+runCLIAction();
